This commit was generated by merge_to_master.py.

+++ chrome/browser/sync_file_system/drive/local_change_processor_delegate.cc
@@ -9,6 +9,7 @@
 #include "chrome/browser/sync_file_system/drive/api_util.h"
 #include "chrome/browser/sync_file_system/drive_file_sync_service.h"
 #include "chrome/browser/sync_file_system/drive_metadata_store.h"
+#include "chrome/browser/sync_file_system/logger.h"
 #include "webkit/browser/fileapi/syncable/syncable_file_system_util.h"
 
 namespace sync_file_system {
@@ -21,6 +22,7 @@ LocalChangeProcessorDelegate::LocalChangeProcessorDelegate(
     const SyncFileMetadata& local_metadata,
     const fileapi::FileSystemURL& url)
     : sync_service_(sync_service),
+      operation_(SYNC_OPERATION_NONE),
       url_(url),
       local_change_(local_change),
       local_path_(local_path),
@@ -34,6 +36,7 @@ LocalChangeProcessorDelegate::~LocalChangeProcessorDelegate() {}
 void LocalChangeProcessorDelegate::Run(const SyncStatusCallback& callback) {
   // TODO(nhiroki): support directory operations (http://crbug.com/161442).
   DCHECK(IsSyncFSDirectoryOperationEnabled() || !local_change_.IsDirectory());
+  operation_ = SYNC_OPERATION_NONE;
 
   has_drive_metadata_ =
       metadata_store()->ReadEntry(url_, &drive_metadata_) == SYNC_STATUS_OK;
@@ -71,46 +74,49 @@ void LocalChangeProcessorDelegate::DidGetOriginRoot(
               drive_metadata_.type())
       : SYNC_FILE_TYPE_UNKNOWN;
 
-  LocalSyncOperationType operation = LocalSyncOperationResolver::Resolve(
+  DCHECK_EQ(SYNC_OPERATION_NONE, operation_);
+  operation_ = LocalSyncOperationResolver::Resolve(
       local_change_,
       has_remote_change_ ? &remote_change_.change : NULL,
       has_drive_metadata_ ? &drive_metadata_ : NULL);
 
-  DVLOG(1) << "ApplyLocalChange for " << url_.DebugString()
-           << " local_change:" << local_change_.DebugString()
-           << " ==> operation:" << operation;
+  util::Log(logging::LOG_VERBOSE, FROM_HERE,
+            "ApplyLocalChange for %s local_change:%s ===> %s",
+            url_.DebugString().c_str(),
+            local_change_.DebugString().c_str(),
+            SyncOperationTypeToString(operation_));
 
-  switch (operation) {
-    case LOCAL_SYNC_OPERATION_ADD_FILE:
+  switch (operation_) {
+    case SYNC_OPERATION_ADD_FILE:
       UploadNewFile(callback);
       return;
-    case LOCAL_SYNC_OPERATION_ADD_DIRECTORY:
+    case SYNC_OPERATION_ADD_DIRECTORY:
       CreateDirectory(callback);
       return;
-    case LOCAL_SYNC_OPERATION_UPDATE_FILE:
+    case SYNC_OPERATION_UPDATE_FILE:
       UploadExistingFile(callback);
       return;
-    case LOCAL_SYNC_OPERATION_DELETE:
+    case SYNC_OPERATION_DELETE:
       Delete(callback);
       return;
-    case LOCAL_SYNC_OPERATION_NONE:
+    case SYNC_OPERATION_NONE:
       callback.Run(SYNC_STATUS_OK);
       return;
-    case LOCAL_SYNC_OPERATION_CONFLICT:
+    case SYNC_OPERATION_CONFLICT:
       HandleConflict(callback);
       return;
-    case LOCAL_SYNC_OPERATION_RESOLVE_TO_LOCAL:
+    case SYNC_OPERATION_RESOLVE_TO_LOCAL:
       ResolveToLocal(callback);
       return;
-    case LOCAL_SYNC_OPERATION_RESOLVE_TO_REMOTE:
+    case SYNC_OPERATION_RESOLVE_TO_REMOTE:
       ResolveToRemote(callback, remote_file_type);
       return;
-    case LOCAL_SYNC_OPERATION_DELETE_METADATA:
+    case SYNC_OPERATION_DELETE_METADATA:
       DeleteMetadata(base::Bind(
           &LocalChangeProcessorDelegate::DidApplyLocalChange,
           weak_factory_.GetWeakPtr(), callback, google_apis::HTTP_SUCCESS));
       return;
-    case LOCAL_SYNC_OPERATION_FAIL: {
+    case SYNC_OPERATION_FAIL: {
       callback.Run(SYNC_STATUS_FAILED);
       return;
     }
@@ -197,6 +203,11 @@ void LocalChangeProcessorDelegate::DidCreateDirectory(
 void LocalChangeProcessorDelegate::UploadExistingFile(
     const SyncStatusCallback& callback) {
   DCHECK(has_drive_metadata_);
+  if (drive_metadata_.resource_id().empty()) {
+    UploadNewFile(callback);
+    return;
+  }
+
   api_util()->UploadExistingFile(
       drive_metadata_.resource_id(),
       drive_metadata_.md5_checksum(),
@@ -223,19 +234,16 @@ void LocalChangeProcessorDelegate::DidUploadExistingFile(
           SYNC_ACTION_UPDATED,
           SYNC_DIRECTION_LOCAL_TO_REMOTE);
       return;
-    case google_apis::HTTP_CONFLICT: {
+    case google_apis::HTTP_CONFLICT:
       HandleConflict(callback);
       return;
-    }
-    case google_apis::HTTP_NOT_MODIFIED: {
+    case google_apis::HTTP_NOT_MODIFIED:
       DidApplyLocalChange(callback,
                           google_apis::HTTP_SUCCESS, SYNC_STATUS_OK);
       return;
-    }
-    case google_apis::HTTP_NOT_FOUND: {
+    case google_apis::HTTP_NOT_FOUND:
       UploadNewFile(callback);
       return;
-    }
     default: {
       const SyncStatusCode status =
           GDataErrorCodeToSyncStatusCodeWrapper(error);
@@ -252,6 +260,12 @@ void LocalChangeProcessorDelegate::Delete(
     callback.Run(SYNC_STATUS_OK);
     return;
   }
+
+  if (drive_metadata_.resource_id().empty()) {
+    DidDelete(callback, google_apis::HTTP_NOT_FOUND);
+    return;
+  }
+
   api_util()->DeleteFile(
       drive_metadata_.resource_id(),
       drive_metadata_.md5_checksum(),
@@ -265,27 +279,29 @@ void LocalChangeProcessorDelegate::DidDelete(
   DCHECK(has_drive_metadata_);
 
   switch (error) {
-    // Regardless of whether the deletion has succeeded (HTTP_SUCCESS) or
-    // has failed with ETag conflict error (HTTP_PRECONDITION or HTTP_CONFLICT)
-    // we should just delete the drive_metadata.
-    // In the former case the file should be just gone now, and
-    // in the latter case the remote change will be applied in a future
-    // remote sync.
     case google_apis::HTTP_SUCCESS:
-    case google_apis::HTTP_PRECONDITION:
-    case google_apis::HTTP_CONFLICT:
+    case google_apis::HTTP_NOT_FOUND:
       DeleteMetadata(base::Bind(
           &LocalChangeProcessorDelegate::DidApplyLocalChange,
-          weak_factory_.GetWeakPtr(), callback, error));
+          weak_factory_.GetWeakPtr(), callback, google_apis::HTTP_SUCCESS));
       sync_service_->NotifyObserversFileStatusChanged(
           url_,
           SYNC_FILE_STATUS_SYNCED,
           SYNC_ACTION_DELETED,
           SYNC_DIRECTION_LOCAL_TO_REMOTE);
       return;
-    case google_apis::HTTP_NOT_FOUND:
-      DidApplyLocalChange(callback,
-                          google_apis::HTTP_SUCCESS, SYNC_STATUS_OK);
+    case google_apis::HTTP_PRECONDITION:
+    case google_apis::HTTP_CONFLICT:
+      // Delete |drive_metadata| on the conflict case.
+      // Conflicted remote change should be applied as a future remote change.
+      DeleteMetadata(base::Bind(
+          &LocalChangeProcessorDelegate::DidDeleteMetadataForDeletionConflict,
+          weak_factory_.GetWeakPtr(), callback));
+      sync_service_->NotifyObserversFileStatusChanged(
+          url_,
+          SYNC_FILE_STATUS_SYNCED,
+          SYNC_ACTION_DELETED,
+          SYNC_DIRECTION_LOCAL_TO_REMOTE);
       return;
     default: {
       const SyncStatusCode status =
@@ -297,8 +313,19 @@ void LocalChangeProcessorDelegate::DidDelete(
   }
 }
 
+void LocalChangeProcessorDelegate::DidDeleteMetadataForDeletionConflict(
+    const SyncStatusCallback& callback,
+    SyncStatusCode status) {
+  callback.Run(SYNC_STATUS_OK);
+}
+
 void LocalChangeProcessorDelegate::ResolveToLocal(
     const SyncStatusCallback& callback) {
+  if (drive_metadata_.resource_id().empty()) {
+    DidDeleteFileToResolveToLocal(callback, google_apis::HTTP_NOT_FOUND);
+    return;
+  }
+
   api_util()->DeleteFile(
       drive_metadata_.resource_id(),
       drive_metadata_.md5_checksum(),
@@ -312,7 +339,6 @@ void LocalChangeProcessorDelegate::DidDeleteFileToResolveToLocal(
     google_apis::GDataErrorCode error) {
   if (error != google_apis::HTTP_SUCCESS &&
       error != google_apis::HTTP_NOT_FOUND) {
-    remote_change_handler()->RemoveChangeForURL(url_);
     callback.Run(GDataErrorCodeToSyncStatusCodeWrapper(error));
     return;
   }
@@ -364,6 +390,13 @@ void LocalChangeProcessorDelegate::DidApplyLocalChange(
     const SyncStatusCallback& callback,
     const google_apis::GDataErrorCode error,
     SyncStatusCode status) {
+  if ((operation_ == SYNC_OPERATION_DELETE ||
+       operation_ == SYNC_OPERATION_DELETE_METADATA) &&
+      (status == SYNC_FILE_ERROR_NOT_FOUND ||
+       status == SYNC_DATABASE_ERROR_NOT_FOUND)) {
+    status = SYNC_STATUS_OK;
+  }
+
   if (status == SYNC_STATUS_OK) {
     remote_change_handler()->RemoveChangeForURL(url_);
     status = GDataErrorCodeToSyncStatusCodeWrapper(error);
@@ -502,8 +535,9 @@ void LocalChangeProcessorDelegate::HandleManualResolutionCase(
 
 void LocalChangeProcessorDelegate::HandleLocalWinCase(
     const SyncStatusCallback& callback) {
-  DVLOG(1) << "Resolving conflict for local sync:"
-           << url_.DebugString() << ": LOCAL WIN";
+  util::Log(logging::LOG_VERBOSE, FROM_HERE,
+            "Resolving conflict for local sync: %s: LOCAL WIN",
+            url_.DebugString().c_str());
 
   DCHECK(!drive_metadata_.resource_id().empty());
   if (!has_drive_metadata_) {
@@ -518,8 +552,9 @@ void LocalChangeProcessorDelegate::HandleLocalWinCase(
 void LocalChangeProcessorDelegate::HandleRemoteWinCase(
     const SyncStatusCallback& callback,
     SyncFileType remote_file_type) {
-  DVLOG(1) << "Resolving conflict for local sync:"
-           << url_.DebugString() << ": REMOTE WIN";
+  util::Log(logging::LOG_VERBOSE, FROM_HERE,
+            "Resolving conflict for local sync: %s: REMOTE WIN",
+            url_.DebugString().c_str());
   ResolveToRemote(callback, remote_file_type);
 }
